modified prod and staging data, wrote basic functions
The production1 and production2 are mock data respectively for 2 seperate requests
with the same parameters in production environment.

The staging data is the data received after we pass the same parameters as above to the
staging/testing environment.

We compare the 2 production objects to get paths of key values that will change irrespective of
the code change or refactoring. Things like, date, time etc...

The paths are then passed to another function along with the production data and staging data.
The continuously changing keys are removed using paths from both prod and staging,
then we check if there is any differece in the objects.

If there is a difference, then something is not good with the refctoring,
if there isnt a differece, then refactoring went well